feat(eslint-plugin): deprecate current rules
