+++ README.md
@@ -1,4 +1,4 @@
-# Tinkoff eslint configurations
+# Tinkoff linter configurations
 
 Monorepo for all the tooling which enables ESLint, Prettier, Stylelint to support TypeScript projects.
