+++ package.json
@@ -40,8 +40,8 @@
         "@types/eslint": "^7.2.0",
         "@types/jest": "^26.0.0",
         "@types/node": "^14.0.1",
-        "@typescript-eslint/eslint-plugin": "4.32.0",
-        "@typescript-eslint/parser": "4.32.0",
+        "@typescript-eslint/eslint-plugin": "^5.3.0",
+        "@typescript-eslint/parser": "^5.3.0",
         "babel-jest": "^26.0.1",
         "concurrently": "5.0.1",
         "conventional-changelog-conventionalcommits": "^4.2.3",
