+++ packages/eslint-config-angular/package.json
@@ -22,7 +22,7 @@
         "@tinkoff/eslint-config": "^1.5.0"
     },
     "peerDependencies": {
-        "@typescript-eslint/eslint-plugin": "4.32.0",
+        "@typescript-eslint/eslint-plugin": "^5.3.0",
         "eslint": "^7.5.0",
         "eslint-plugin-import": "^2.22.0",
         "eslint-plugin-sort-class-members": "^1.7.0"
