+++ packages/eslint-config/package.json
@@ -21,8 +21,8 @@
     "dependencies": {
         "@babel/eslint-parser": "^7.14.7",
         "@babel/eslint-plugin": "^7.14.5",
-        "@typescript-eslint/eslint-plugin": "4.32.0",
-        "@typescript-eslint/parser": "4.32.0",
+        "@typescript-eslint/eslint-plugin": "^5.3.0",
+        "@typescript-eslint/parser": "^5.3.0",
         "eslint": "^7.5.0",
         "eslint-config-airbnb-base": "^14.2.0",
         "eslint-config-prettier": "^6.11.0",
