+++ packages/eslint-plugin/README.md
@@ -1,0 +1,18 @@
+### Usage
+
+```json5
+{
+    plugins: ['@tinkoff/eslint-plugin'],
+    rules: {
+        // ...
+    },
+}
+```
+
+### Rules
+
+| Name                         | Description                                                  | Deprecated |
+| ---------------------------- | ------------------------------------------------------------ | ---------- |
+| -                            | -                                                            | -          |
+| @tinkoff/member-ordering     | Require a consistent member declaration order                | yes        |
+| @tinkoff/require-return-type | Require explicit return types on functions and class methods | yes        |
