+++ packages/eslint-plugin/lib/index.js
@@ -1,6 +1,6 @@
 module.exports = {
     rules: {
-        'member-ordering': require('./member-ordering'),
-        'require-return-type': require('./require-return-type'),
+        'member-ordering': require('./deprecated/member-ordering'),
+        'require-return-type': require('./deprecated/require-return-type'),
     },
 };
