+++ packages/eslint-plugin/package.json
@@ -14,8 +14,8 @@
         "url": "https://github.com/TinkoffCreditSystems/linters.git"
     },
     "dependencies": {
-        "@typescript-eslint/eslint-plugin": "4.32.0",
-        "@typescript-eslint/experimental-utils": "4.32.0",
+        "@typescript-eslint/eslint-plugin": "^5.3.0",
+        "@typescript-eslint/experimental-utils": "^5.3.0",
         "lodash": "^4.17.15"
     },
     "peerDependencies": {
