+++ packages/linters/package.json
@@ -17,10 +17,10 @@
     },
     "dependencies": {
         "@tinkoff/eslint-plugin": "^1.7.1",
-        "@typescript-eslint/eslint-plugin": "4.32.0",
-        "@typescript-eslint/eslint-plugin-tslint": "4.32.0",
-        "@typescript-eslint/experimental-utils": "4.32.0",
-        "@typescript-eslint/parser": "4.32.0",
+        "@typescript-eslint/eslint-plugin": "^5.3.0",
+        "@typescript-eslint/eslint-plugin-tslint": "^5.3.0",
+        "@typescript-eslint/experimental-utils": "^5.3.0",
+        "@typescript-eslint/parser": "^5.3.0",
         "eslint": "^7.5.0",
         "eslint-config-prettier": "^6.11.0",
         "eslint-plugin-functional": "^3.7.2",
